+++ tamr_unify_client/models/project/resource.py
@@ -47,8 +47,7 @@ class Project(BaseResource):
         from tamr_unify_client.models.attribute.collection import AttributeCollection
 
         alias = self.api_path + "/attributes"
-        resource_json = self.client.get(alias).successful().json()
-        return AttributeCollection.from_json(self.client, resource_json, alias)
+        return AttributeCollection(self.client, alias)
 
     def unified_dataset(self):
         """Unified dataset for this project.